+++ decrypt.go
@@ -26,7 +26,7 @@ func Decrypt(w io.Writer, clearHashes, cipherChunks ChunkStore, key [32]byte) er
 	if err != nil {
 		return errors.Wrap(err, "counting clear hashes")
 	}
-	for index := uint64(0); index < uint64(nhashes); index++ {
+	for index := int64(0); index < nhashes; index++ {
 		wantClearHash, err := clearHashes.Get(index)
 		if err != nil {
 			return errors.Wrapf(err, "getting clear hash %d", index)
@@ -38,7 +38,7 @@ func Decrypt(w io.Writer, clearHashes, cipherChunks ChunkStore, key [32]byte) er
 		}
 		Crypt(key, chunk, index)
 
-		m := binary.PutUvarint(chunkWithPrefix[:], index)
+		m := binary.PutUvarint(chunkWithPrefix[:], uint64(index))
 		copy(chunkWithPrefix[m:], chunk)
 
 		merkle.LeafHash(hasher, gotClearHash[:0], chunkWithPrefix[m:m+len(chunk)])
@@ -58,7 +58,7 @@ func Decrypt(w io.Writer, clearHashes, cipherChunks ChunkStore, key [32]byte) er
 // BadClearHashError gives the index of a cleartext chunk whose hash doesn't have the expected value.
 type BadClearHashError struct {
 	// Index is the index of the chunk and of the hash within their respective ChunkStores.
-	Index uint64
+	Index int64
 }
 
 func (e BadClearHashError) Error() string {
